+++ oxapi/__init__.py
@@ -4,15 +4,12 @@ import os
 base_url = None
 api_key = None
 
-__version__: str = version
-log_level = log_level
+__version__ = "1.0.0"
+log_level = os.getenv("LOG_LEVEL", "INFO")
 logger = logging.getLogger(name="OxAPI Logger")
 logger.setLevel(level=log_level)
 
-default_api_version: str = default_api_version
-default_model_version: str = default_model_version
-
-os.environ["GEVENT_SUPPORT"] = True
+os.environ["GEVENT_SUPPORT"] = "True"
 
 try:
     api_key = os.environ["OXAPI_KEY"]
@@ -22,9 +19,12 @@ except KeyError:
     )
 
 from oxapi.asynch import AsyncCallPipe
-from oxapi.config import default_api_version, default_model_version, log_level, version
+from oxapi.config import default_api_version, default_model_version
 from oxapi.nlp.classification import Classification
 from oxapi.nlp.completion import Completion
 from oxapi.nlp.encoding import Encoding
 from oxapi.nlp.pipeline import Pipeline
 from oxapi.nlp.transformation import Transformation
+
+default_api_version: str = default_api_version
+default_model_version: str = default_model_version