+++ oxapi/config.py
@@ -1,4 +1,2 @@
-log_level = "DEBUG"
-version = "1.0.0"
 default_api_version = "v1"
 default_model_version = "v1"